Implementation Fixed Point Square Root Algorithm with Pipe-line
For given 32-bit input (right 16 bits for decimal, left 16 bits for fraction), the square root (output sonuc) will be 16 bits but I completed it with zeros to 64 bits to connect with my top module.

The pipe-line's depth is 16 for 32-bit input.

circuitWithoutPipeline.png includes the circuit that I implemented in Vivado.